chore(compat-corpus): dedupe readIf/firstDiffLine across verify/cluster scripts
readIf (existsSync-then-read) and firstDiffLine (first differing line,
truncated for display) were reimplemented byte-for-byte in verify.mjs,
svelte2tsx-verify.mjs, fmt-verify.mjs, cluster.mjs, and
svelte2tsx-cluster.mjs. Move both into normalize.mjs (already the shared
module for stripBlankLines/flattenTemplateHoles/astEquivalent) and import
from there instead. Purely mechanical: no logic changed, verified with a
synthetic before/after corpus fixture producing byte-identical
verify/cluster output and report JSON (aside from the generatedAt
timestamp) on both sides of the refactor.

Findings-09 P2: "verify.mjs / svelte2tsx-verify.mjs / fmt-verify.mjs で
ラチェット判定・firstDiffLine・readIf が三重に重複" and "cluster.mjs と
svelte2tsx-cluster.mjs の diffSignature / クラスタ集計ロジックが完全重複"
(readIf portion only — diffSignature/add and the deeper baseline/
regression-computation blocks are left alone; the differing sample/ids-cap
behavior between the two cluster scripts made that riskier to unify
without a full corpus run to validate against).
